third_party/blink/web_tests/platform/generic/fast/forms/select/listbox-selection-expected.txt

<select> selection test for mouse events and keyevents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

1) Select one item with mouse (no previous selection)
PASS selectionPattern("sl1") is "10000"
2) Select one item with mouse (with previous selection)
PASS selectionPattern("sl2") is "10000"
3) Select one item with the keyboard (no previous selection)
PASS selectionPattern("sl3") is "00001"
4) Select one item with the keyboard (with previous selection)
PASS selectionPattern("sl4") is "00100"
5) Attempt to select an item cmd-clicking
PASS selectionPattern("sl5") is "01000"
6) Attempt to select a range shift-clicking
PASS selectionPattern("sl6") is "01000"
7) Attempt to select a range with the keyboard
PASS selectionPattern("sl7") is "00100"
8) Select one item with mouse (no previous selection)
PASS selectionPattern("sl8") is "10000"
9) Select one item with mouse (with previous selection)
PASS selectionPattern("sl9") is "10000"
10) Select one item with the keyboard (no previous selection)
PASS selectionPattern("sl10") is "00001"
11) Select one item with the keyboard (with previous selection)
PASS selectionPattern("sl11") is "00100"
12) Select an item cmd-clicking
PASS selectionPattern("sl12") is "01100"
13) Select a range shift-clicking
PASS selectionPattern("sl13") is "11110"
14) Select a range with the keyboard
PASS selectionPattern("sl14") is "01100"
15) Drag upside-down
PASS selectionPattern("sl15") is "01000"
PASS mouseMoveToOption("sl15", 3); selectionPattern("sl15") is "01110"
PASS mouseMoveToOption("sl15", 2); selectionPattern("sl15") is "01100"
PASS sl15.add(new Option("a")); eventSender.keyDown("ArrowUp"); selectionPattern("sl15") is "010000"
16) Active-selection after type-ahead
PASS mouseClickOnSelect('sl16', 1); selectionPattern('sl16') is "01000"
PASS keyDownOnSelect('sl16', 'e'); selectionPattern('sl16') is "00001"
PASS keyDownOnSelect('sl16', 'ArrowUp', 'rangeSelectionKey'); selectionPattern('sl16') is "00011"
PASS successfullyParsed is true

TEST COMPLETE
